requirements2.txt
pdf2image==1.16.3
Pillow==10.1.0
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
